requirements.txt
PySastrawi==1.2.0